requirements.txt
# ===================================
# Sistema de Reconocimiento Facial
# Backend con Django + DRF + JWT + Authz
# ===================================

# === Core Django Framework ===
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.6.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# === JWT Authentication ===
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# === API Documentation ===
drf-spectacular==0.28.0
PyYAML==6.0.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
uritemplate==4.2.0

# === Environment Management ===
python-dotenv==1.1.1

# === Image Processing & Face Recognition ===
Pillow==11.3.0
numpy>=1.22.2
opencv-python==4.12.0.88

# === Face Recognition Library ===
cmake==4.1.0
dlib>=19.24.0
face-recognition>=1.3.0

# === Advanced AI & Computer Vision (Optional) ===
# YOLO para detección avanzada de caras
ultralytics==8.0.196

# PyTorch para modelos de IA (requerido por YOLO)
torch>=2.0.0
torchvision>=0.15.0

# MediaPipe para detección facial alternativa
# mediapipe==0.10.21  # Comentado temporalmente por conflictos de dependencias

# TensorFlow (opcional para modelos custom)
# tensorflow==2.13.0  # Descomenta si necesitas TensorFlow

# === Dropbox SDK ===
dropbox>=11.36.2

# === Azure Cognitive Services (Optional) ===
azure-cognitiveservices-vision-face==0.6.1
azure-common==1.1.28
azure-core==1.35.1
azure-mgmt-core==1.6.0
msrest==0.7.1
# === OCR ===
pytesseract==0.3.13

# === HTTP & Requests ===
requests==2.32.5
requests-oauthlib==2.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
urllib3==2.5.0
idna==3.10

# === OAuth & Security ===
oauthlib==3.3.1

# === Development & Testing ===
responses==0.25.8

# === Data Structures & Validation ===
attrs==25.3.0
referencing==0.36.2
rpds-py==0.27.1
typing_extensions==4.15.0
isodate==0.7.2
six==1.17.0

# === Additional Dependencies for Production ===
# Uncomment for production deployment
# gunicorn>=21.2.0
# psycopg2-binary>=2.9.0  # For PostgreSQL
# redis>=5.0.0  # For caching
# celery>=5.3.0  # For background tasks
# whitenoise>=6.5.0  # For static files

# === Development Tools (Optional) ===
# Uncomment for development
# django-debug-toolbar>=4.2.0
# django-extensions>=3.2.0
# black>=23.0.0  # Code formatting
# flake8>=6.0.0  # Code linting
# pytest>=7.4.0  # Testing
# pytest-django>=4.5.0  # Django testing
